# apigen/example/internal/api/types.gen.go
// Code generated by apigen standalone schema model emitter. DO NOT EDIT.
package api

type CreateTodoRequest struct {
	Title string `json:"title"`
}

type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

type ListTodosResponse struct {
	Data []Todo `json:"data"`
}

type Todo struct {
	Id     string `json:"id"`
	Status string `json:"status"`
	Title  string `json:"title"`
}

type GenericRequest struct {
	Payload *map[string]string `json:"payload,omitempty"`
}

type GenericResponse struct {
	Data *map[string]string `json:"data,omitempty"`
}

type ListTodosParams = GenListTodosParams

type CreateTodoJSONRequestBody = GenSchemaCreateTodoRequest